debug

+++ main/src/main/scala/org/clulab/polarity/ml/DeepLearningPolarityClassifier.scala
@@ -235,7 +235,7 @@ class DeepLearningPolarityClassifier() extends PolarityClassifier{
 
     println("==========================================")
     println(lemmas.slice(start, end).toList)
-    println(y_pred.value())
+    println(y_pred.value().toFloat())
     //    scala.io.StdIn.readLine()
 
     if (y_pred.value().toFloat>0.5){
